# apps/api/openapi/spec.yaml
openapi: 3.0.0
info:
  title: lomi.africa Payments API
  version: 1.0.0
  description: API of lomi. payment orchestration platform
servers:
  - url: https://api.lomi.africa/v1
    
# API Schema to Database Table Mapping:
# - Merchant (API) -> merchants (DB)
# - Product (API) -> merchant_products (DB)
# - SubscriptionPlan (API) -> subscription_plans (DB)
# - Subscription (API) -> merchant_subscriptions (DB)
# - Transaction (API) -> transactions (DB)
# - Customer (API) -> customers (DB)
# - PaymentLink (API) -> payment_links (DB)
# - Refund (API) -> refunds (DB)
# - Webhook (API) -> webhooks (DB)
# - CheckoutSession (API) -> checkout_sessions (DB)
# - MerchantAccount (API) -> merchant_accounts (DB)
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-KEY
  responses:
    UnauthorizedError:
      description: API key is missing or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ValidationError:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFoundError:
      description: The requested resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
...
    Subscription:
      type: object
      properties:
        subscription_id:
          type: string
          format: uuid
        merchant_id:
          type: string
          format: uuid
        organization_id:
          type: string
          format: uuid
        plan_id:
          type: string
          format: uuid
        customer_id:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/SubscriptionStatus'
        start_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
        next_billing_date:
          type: string
          format: date
        metadata:
          type: object
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - subscription_id
        - merchant_id
        - organization_id
        - plan_id
        - customer_id
        - status
        - start_date
...
    SubscriptionStatus:
      type: string
      enum:
        - pending
        - active
        - paused
        - cancelled
        - expired
        - past_due
        - trial
...
security:
  - ApiKeyAuth: []
paths:
...
  /customer-subscriptions:
    get:
      summary: List customer subscriptions
      description: List all subscriptions for a merchant
      operationId: listCustomerSubscriptions
      parameters:
        - name: merchant_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: customer_id
          in: query
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          schema:
            type: string
            enum:
              - pending
              - active
              - paused
              - cancelled
              - expired
              - past_due
              - trial
      responses:
        '200':
          description: List of subscriptions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Subscription'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /customer-subscriptions/{subscription_id}:
    get:
      summary: Get subscription details
      description: Get details of a specific subscription
      operationId: getSubscription
      parameters:
        - name: subscription_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: merchant_id
          in: query
          required: true
          description: The ID of the merchant owning the subscription.
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Subscription details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subscription'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
          
    patch:
      summary: Update subscription
      description: Update subscription details
      operationId: updateSubscription
      parameters:
        - name: subscription_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: merchant_id
          in: query
          required: true
          description: The ID of the merchant owning the subscription.
          schema:
            type: string
            format: uuid
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum:
                    - active
                    - paused
                    - cancelled
                end_date:
                  type: string
                  format: date
                next_billing_date:
                  type: string
                  format: date
                metadata:
                  type: object
      responses:
        '200':
          description: Subscription updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subscription'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
          
    delete:
      summary: Cancel subscription
      description: Cancel a subscription
      operationId: cancelSubscription
      parameters:
        - name: subscription_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: merchant_id
          in: query
          required: true
          description: The ID of the merchant owning the subscription.
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Subscription cancelled successfully
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
...